--- parcels/_core/constants.py
@@ -1,3 +1,0 @@
-# Used in the searching of 1D arrays for time and space
-LEFT_OUT_OF_BOUNDS = -2
-RIGHT_OUT_OF_BOUNDS = -1
