+++ parcels/_core/xgrid.py
@@ -8,8 +8,7 @@ import xarray as xr
 import xgcm
 
 from parcels._core.basegrid import BaseGrid
-from parcels._core.index_search import _search_indices_curvilinear_2d
-from parcels._core.utils.array import _search_1d_array
+from parcels._core.index_search import _search_1d_array, _search_indices_curvilinear_2d
 from parcels._typing import assert_valid_mesh
 
 _XGRID_AXES = Literal["X", "Y", "Z"]
